+++ app/src/main/java/org/sopt/and/feature/signin/SignInRoute.kt
@@ -1,18 +1,20 @@
-package org.sopt.and.feature.signup
+package org.sopt.and.feature.signin
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Text
+import androidx.compose.material3.VerticalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -21,16 +23,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
-import androidx.navigation.NavHostController
 import org.sopt.and.R
 import org.sopt.and.core.designsystem.component.SocialLoginButton
 import org.sopt.and.core.designsystem.component.textfield.EmailTextField
@@ -39,123 +38,133 @@ import org.sopt.and.core.designsystem.theme.ANDANDROIDTheme
 import org.sopt.and.core.extension.toast
 
 @Composable
-fun SignUpRoute(
-    navigateToSignIn: (String, String) -> Unit,
-    viewModel: SignUpViewModel = hiltViewModel(),
+fun SignInRoute(
+    navigateToSignUp: () -> Unit,
+    navigateToMyPage: (String, String) -> Unit,
+    viewModel: SignInViewModel = hiltViewModel(),
     modifier: Modifier = Modifier,
 ) {
-    val signUpState by viewModel.signUpState.collectAsStateWithLifecycle()
+    val signInState by viewModel.signInState.collectAsStateWithLifecycle()
 
     val context = LocalContext.current
     val lifecycleOwner = LocalLifecycleOwner.current
 
-    LaunchedEffect(viewModel.signUpSideEffect, lifecycleOwner) {
-        viewModel.signUpSideEffect.flowWithLifecycle(lifecycle = lifecycleOwner.lifecycle)
+    LaunchedEffect(viewModel.signInSideEffect, lifecycleOwner) {
+        viewModel.signInSideEffect.flowWithLifecycle(lifecycle = lifecycleOwner.lifecycle)
             .collect { sideEffect ->
                 when (sideEffect) {
-                    is SignUpSideEffect.Toast -> context.toast(sideEffect.message)
-                    is SignUpSideEffect.NavigateToSignIn -> {
-                        context.toast(R.string.sign_up_success)
-                        navigateToSignIn(signUpState.email, signUpState.password)
+                    is SignInSideEffect.ShowToast -> {
+                        context.toast(sideEffect.message)
+                    }
+
+                    is SignInSideEffect.ShowSnackBar -> {}
+                    is SignInSideEffect.NavigateToSignUp -> navigateToSignUp()
+                    is SignInSideEffect.NavigateToMyPage -> {
+                        navigateToMyPage(signInState.email, signInState.password)
                     }
                 }
             }
     }
 
-    SignUpScreen(
-        onSignUpButtonClick = viewModel::onSignUpClick,
+    SignInScreen(
+        onSignUpButtonClick = viewModel::onSignUpButtonClick,
+        onSignInButtonClick = viewModel::onLoginButtonClick,
         onIdChange = viewModel::updateEmail,
         onPasswordChange = viewModel::updatePassword,
+        signInState = signInState,
         modifier = modifier,
-        signUpState = signUpState,
     )
-
 }
 
 @Composable
-fun SignUpScreen(
+fun SignInScreen(
+    onSignInButtonClick: () -> Unit,
     onSignUpButtonClick: () -> Unit,
     onIdChange: (String) -> Unit,
     onPasswordChange: (String) -> Unit,
+    signInState: SignInState,
     modifier: Modifier = Modifier,
-    signUpState: SignUpState,
 ) {
     Column(
         modifier = modifier
             .fillMaxSize()
             .background(Color.Black),
     ) {
-        SignUpTitle(
-            modifier = Modifier
-                .padding(
-                    top = 30.dp,
-                    start = 20.dp,
-                ),
-            whiteText = stringResource(R.string.sign_up_title_email_and_password),
-            greyText = stringResource(R.string.sign_up_title_only),
-        )
-
-        SignUpTitle(
-            modifier = Modifier
-                .padding(start = 20.dp),
-            whiteText = stringResource(R.string.sign_up_title_enjoy_wavve),
-            greyText = stringResource(R.string.sign_up_title_able),
-        )
-
         EmailTextField(
-            email = signUpState.email,
-            hint = stringResource(R.string.sign_up_email_hint),
+            email = signInState.email,
+            hint = stringResource(R.string.sign_in_email),
             onValueChange = onIdChange,
             modifier = Modifier
-                .padding(top = 30.dp)
+                .padding(top = 50.dp)
                 .padding(horizontal = 20.dp),
         )
-
-        Row(
+        PasswordTextField(
+            password = signInState.password,
+            hint = stringResource(R.string.sign_in_password),
+            onValueChange = onPasswordChange,
             modifier = Modifier
-                .padding(top = 10.dp)
+                .padding(top = 5.dp)
                 .padding(horizontal = 20.dp),
-        ) {
-            Icon(
-                painter = painterResource(R.drawable.ic_info_24),
-                contentDescription = stringResource(R.string.img_email),
-                tint = Color.Gray,
+            onDoneAction = onSignInButtonClick
+        )
+
+        Button(
+            onClick = onSignInButtonClick,
+            shape = RoundedCornerShape(20.dp),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 20.dp)
+                .padding(horizontal = 20.dp),
+            colors = ButtonDefaults.buttonColors(
+                containerColor = Color.Blue,
             )
+        ) {
             Text(
-                text = stringResource(R.string.sign_up_email_description),
-                fontSize = 12.sp,
-                color = Color.Gray,
-                modifier = Modifier
-                    .padding(start = 3.dp, end = 10.dp),
+                text = stringResource(R.string.sign_in),
+                color = Color.White,
             )
         }
 
-        PasswordTextField(
-            password = signUpState.password,
-            hint = stringResource(R.string.sign_up_password_hint),
-            onValueChange = onPasswordChange,
-            modifier = Modifier
-                .padding(top = 10.dp)
-                .padding(horizontal = 20.dp),
-            onDoneAction = onSignUpButtonClick,
-        )
-
         Row(
+            horizontalArrangement = Arrangement.Center,
+            verticalAlignment = Alignment.CenterVertically,
             modifier = Modifier
-                .padding(top = 10.dp)
-                .padding(horizontal = 20.dp),
+                .fillMaxWidth()
+                .padding(top = 15.dp),
         ) {
-            Icon(
-                painter = painterResource(R.drawable.ic_info_24),
-                contentDescription = stringResource(R.string.img_password),
-                tint = Color.Gray,
-            )
             Text(
-                text = stringResource(R.string.sign_up_password_description),
+                text = stringResource(R.string.sign_in_find_id),
                 fontSize = 12.sp,
                 color = Color.Gray,
                 modifier = Modifier
-                    .padding(start = 3.dp, end = 10.dp),
+                    .padding(horizontal = 8.dp),
+            )
+            VerticalDivider(
+                modifier = Modifier
+                    .height(10.dp)
+                    .background(Color.Gray)
+            )
+            Text(
+                text = stringResource(R.string.sign_in_change_password),
+                fontSize = 12.sp,
+                color = Color.Gray,
+                modifier = Modifier
+                    .padding(horizontal = 8.dp),
+            )
+            VerticalDivider(
+                modifier = Modifier
+                    .height(10.dp)
+                    .background(Color.Gray)
+            )
+            Text(
+                text = stringResource(R.string.sign_up_main_title),
+                fontSize = 12.sp,
+                color = Color.Gray,
+                modifier = Modifier
+                    .padding(horizontal = 8.dp)
+                    .clickable {
+                        onSignUpButtonClick()
+                    },
             )
         }
 
@@ -201,46 +210,6 @@ fun SignUpScreen(
                     end = 20.dp
                 ),
         )
-
-        Spacer(
-            modifier = Modifier.weight(1f)
-        )
-
-        Text(
-            text = stringResource(R.string.sign_up_button),
-            fontSize = 16.sp,
-            color = Color.White,
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .background(Color.Gray)
-                .fillMaxWidth()
-                .clickable {
-                    onSignUpButtonClick()
-                }
-                .padding(vertical = 10.dp)
-        )
-    }
-}
-
-@Composable
-private fun SignUpTitle(
-    modifier: Modifier,
-    whiteText: String,
-    greyText: String,
-) {
-    Row(
-        modifier = modifier,
-    ) {
-        Text(
-            text = whiteText,
-            fontSize = 20.sp,
-            color = Color.White,
-        )
-        Text(
-            text = greyText,
-            fontSize = 20.sp,
-            color = Color.Gray,
-        )
     }
 }
 
@@ -248,12 +217,13 @@ private fun SignUpTitle(
 @Composable
 fun SignInPreview() {
     ANDANDROIDTheme {
-        SignUpScreen(
+        SignInScreen(
             onSignUpButtonClick = { },
-            modifier = Modifier,
+            onSignInButtonClick = { },
             onIdChange = { },
             onPasswordChange = { },
-            signUpState = SignUpState()
+            signInState = SignInState(),
+            modifier = Modifier,
         )
     }
 }
