+++ app/src/main/java/org/sopt/and/feature/signup/SignUpRoute.kt
@@ -1,21 +1,17 @@
-package org.sopt.and.feature.signin
+package org.sopt.and.feature.signup
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
-import androidx.compose.material3.VerticalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -24,14 +20,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
-import androidx.navigation.NavHostController
 import org.sopt.and.R
 import org.sopt.and.core.designsystem.component.SocialLoginButton
 import org.sopt.and.core.designsystem.component.textfield.EmailTextField
@@ -40,133 +37,123 @@ import org.sopt.and.core.designsystem.theme.ANDANDROIDTheme
 import org.sopt.and.core.extension.toast
 
 @Composable
-fun SignInRoute(
-    navigateToSignUp: () -> Unit,
-    navigateToMyPage: (String, String) -> Unit,
-    viewModel: SignInViewModel = hiltViewModel(),
+fun SignUpRoute(
+    navigateToSignIn: (String, String) -> Unit,
+    viewModel: SignUpViewModel = hiltViewModel(),
     modifier: Modifier = Modifier,
 ) {
-    val signInState by viewModel.signInState.collectAsStateWithLifecycle()
+    val signUpState by viewModel.signUpState.collectAsStateWithLifecycle()
 
     val context = LocalContext.current
     val lifecycleOwner = LocalLifecycleOwner.current
 
-    LaunchedEffect(viewModel.signInSideEffect, lifecycleOwner) {
-        viewModel.signInSideEffect.flowWithLifecycle(lifecycle = lifecycleOwner.lifecycle)
+    LaunchedEffect(viewModel.signUpSideEffect, lifecycleOwner) {
+        viewModel.signUpSideEffect.flowWithLifecycle(lifecycle = lifecycleOwner.lifecycle)
             .collect { sideEffect ->
                 when (sideEffect) {
-                    is SignInSideEffect.ShowToast -> {
-                        context.toast(sideEffect.message)
-                    }
-
-                    is SignInSideEffect.ShowSnackBar -> {}
-                    is SignInSideEffect.NavigateToSignUp -> navigateToSignUp()
-                    is SignInSideEffect.NavigateToMyPage -> {
-                        navigateToMyPage(signInState.email, signInState.password)
+                    is SignUpSideEffect.Toast -> context.toast(sideEffect.message)
+                    is SignUpSideEffect.NavigateToSignIn -> {
+                        context.toast(R.string.sign_up_success)
+                        navigateToSignIn(signUpState.email, signUpState.password)
                     }
                 }
             }
     }
 
-    SignInScreen(
-        onSignUpButtonClick = viewModel::onSignUpButtonClick,
-        onSignInButtonClick = viewModel::onLoginButtonClick,
+    SignUpScreen(
+        onSignUpButtonClick = viewModel::onSignUpClick,
         onIdChange = viewModel::updateEmail,
         onPasswordChange = viewModel::updatePassword,
-        signInState = signInState,
         modifier = modifier,
+        signUpState = signUpState,
     )
+
 }
 
 @Composable
-fun SignInScreen(
-    onSignInButtonClick: () -> Unit,
+fun SignUpScreen(
     onSignUpButtonClick: () -> Unit,
     onIdChange: (String) -> Unit,
     onPasswordChange: (String) -> Unit,
-    signInState: SignInState,
     modifier: Modifier = Modifier,
+    signUpState: SignUpState,
 ) {
     Column(
         modifier = modifier
             .fillMaxSize()
             .background(Color.Black),
     ) {
-        EmailTextField(
-            email = signInState.email,
-            hint = stringResource(R.string.sign_in_email),
-            onValueChange = onIdChange,
+        SignUpTitle(
             modifier = Modifier
-                .padding(top = 50.dp)
-                .padding(horizontal = 20.dp),
-        )
-        PasswordTextField(
-            password = signInState.password,
-            hint = stringResource(R.string.sign_in_password),
-            onValueChange = onPasswordChange,
-            modifier = Modifier
-                .padding(top = 5.dp)
-                .padding(horizontal = 20.dp),
-            onDoneAction = onSignInButtonClick
+                .padding(
+                    top = 30.dp,
+                    start = 20.dp,
+                ),
+            whiteText = stringResource(R.string.sign_up_title_email_and_password),
+            greyText = stringResource(R.string.sign_up_title_only),
         )
 
-        Button(
-            onClick = onSignInButtonClick,
-            shape = RoundedCornerShape(20.dp),
+        SignUpTitle(
             modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 20.dp)
+                .padding(start = 20.dp),
+            whiteText = stringResource(R.string.sign_up_title_enjoy_wavve),
+            greyText = stringResource(R.string.sign_up_title_able),
+        )
+
+        EmailTextField(
+            email = signUpState.email,
+            hint = stringResource(R.string.sign_up_email_hint),
+            onValueChange = onIdChange,
+            modifier = Modifier
+                .padding(top = 30.dp)
                 .padding(horizontal = 20.dp),
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Color.Blue,
-            )
+        )
+
+        Row(
+            modifier = Modifier
+                .padding(top = 10.dp)
+                .padding(horizontal = 20.dp),
         ) {
+            Icon(
+                painter = painterResource(R.drawable.ic_info_24),
+                contentDescription = stringResource(R.string.img_email),
+                tint = Color.Gray,
+            )
             Text(
-                text = stringResource(R.string.sign_in),
-                color = Color.White,
+                text = stringResource(R.string.sign_up_email_description),
+                fontSize = 12.sp,
+                color = Color.Gray,
+                modifier = Modifier
+                    .padding(start = 3.dp, end = 10.dp),
             )
         }
 
-        Row(
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically,
+        PasswordTextField(
+            password = signUpState.password,
+            hint = stringResource(R.string.sign_up_password_hint),
+            onValueChange = onPasswordChange,
             modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 15.dp),
+                .padding(top = 10.dp)
+                .padding(horizontal = 20.dp),
+            onDoneAction = onSignUpButtonClick,
+        )
+
+        Row(
+            modifier = Modifier
+                .padding(top = 10.dp)
+                .padding(horizontal = 20.dp),
         ) {
+            Icon(
+                painter = painterResource(R.drawable.ic_info_24),
+                contentDescription = stringResource(R.string.img_password),
+                tint = Color.Gray,
+            )
             Text(
-                text = stringResource(R.string.sign_in_find_id),
+                text = stringResource(R.string.sign_up_password_description),
                 fontSize = 12.sp,
                 color = Color.Gray,
                 modifier = Modifier
-                    .padding(horizontal = 8.dp),
-            )
-            VerticalDivider(
-                modifier = Modifier
-                    .height(10.dp)
-                    .background(Color.Gray)
-            )
-            Text(
-                text = stringResource(R.string.sign_in_change_password),
-                fontSize = 12.sp,
-                color = Color.Gray,
-                modifier = Modifier
-                    .padding(horizontal = 8.dp),
-            )
-            VerticalDivider(
-                modifier = Modifier
-                    .height(10.dp)
-                    .background(Color.Gray)
-            )
-            Text(
-                text = stringResource(R.string.sign_up_main_title),
-                fontSize = 12.sp,
-                color = Color.Gray,
-                modifier = Modifier
-                    .padding(horizontal = 8.dp)
-                    .clickable {
-                        onSignUpButtonClick()
-                    },
+                    .padding(start = 3.dp, end = 10.dp),
             )
         }
 
@@ -212,6 +199,46 @@ fun SignInScreen(
                     end = 20.dp
                 ),
         )
+
+        Spacer(
+            modifier = Modifier.weight(1f)
+        )
+
+        Text(
+            text = stringResource(R.string.sign_up_button),
+            fontSize = 16.sp,
+            color = Color.White,
+            textAlign = TextAlign.Center,
+            modifier = Modifier
+                .background(Color.Gray)
+                .fillMaxWidth()
+                .clickable {
+                    onSignUpButtonClick()
+                }
+                .padding(vertical = 10.dp)
+        )
+    }
+}
+
+@Composable
+private fun SignUpTitle(
+    modifier: Modifier,
+    whiteText: String,
+    greyText: String,
+) {
+    Row(
+        modifier = modifier,
+    ) {
+        Text(
+            text = whiteText,
+            fontSize = 20.sp,
+            color = Color.White,
+        )
+        Text(
+            text = greyText,
+            fontSize = 20.sp,
+            color = Color.Gray,
+        )
     }
 }
 
@@ -219,13 +246,12 @@ fun SignInScreen(
 @Composable
 fun SignInPreview() {
     ANDANDROIDTheme {
-        SignInScreen(
+        SignUpScreen(
             onSignUpButtonClick = { },
-            onSignInButtonClick = { },
+            modifier = Modifier,
             onIdChange = { },
             onPasswordChange = { },
-            signInState = SignInState(),
-            modifier = Modifier,
+            signUpState = SignUpState()
         )
     }
 }
